requirements.txt
antiorm==1.2.1
click==8.1.3
db==0.1.1
db-sqlite3==0.0.1
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyJWT==2.6.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.14.0
